test: fix tests with linebreaks - in prog
tags: test, fix

Why?
tests failed for the place of examples in .R

What?
try to fix how example_pos_start and example_pos_end are computed from parse_fun(), they are then used in add_fun_code_examples()

in progress: example is fixed for 2nd and 3rd function but not for 1st with no

Issue #142